Mamoru

+++ modules/opla-core/src/main/java/br/ufpr/dinf/gres/core/jmetal4/util/MathUtils.java
@@ -17,7 +17,6 @@ import java.util.stream.Collectors;
  * @author elf
  */
 public class MathUtils {
-//
 
     public static HashMap<String, List<List<Double>>> normalize(HashMap<String, List<List<Double>>> listObjectivesValues, String[] objectives) {
 
@@ -74,7 +73,7 @@ public class MathUtils {
         return resultFront.getSolutionSet().stream().map(r -> {
             List<Double> values = new ArrayList<>();
             for (int i = 0; i < r.numberOfObjectives(); i++) {
-                values.add(i, (maxs.get(i) - mins.get(i)) == 0 ? 0 : (r.getObjective(i) - mins.get(i)) / (maxs.get(i) - mins.get(i)));
+                values.add(i, (double)((maxs.get(i) - mins.get(i)) == 0 ? 0 : (r.getObjective(i) - mins.get(i)) / (maxs.get(i) - mins.get(i))));
             }
             return values;
         }).collect(Collectors.toList());
@@ -123,12 +122,6 @@ public class MathUtils {
         return total / values.size();
     }
 
-    /**
-     * Desvio Padrão.
-     *
-     * @param values
-     * @return
-     */
     public static double stDev(List<Double> values) {
         double media = mean(values);
         double somatorio = 0d;
